refactor: dummy data 저장하는 dummy 패키지 생성
