+++ src/resource/dummy/faqList.js
